Minor grammar fixes
A bit unsure about: 
"There is not the same granularity as in the json and yaml specs. e.g. prefix can only exist on the top level." 
vs
"There are not the same granularity as in the json and yaml specs. e.g. prefix can only exist on the top level."

"There is" should be used when the noun is singular, in this case, granularity is singular so it should be right... I think :)